pyproject.toml
[project]
name = "downsyndrome"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask==3.0.3",
    "flask-cors==4.0.1",
    "kagglehub==0.3.4",
    "locust==2.31.6",
    "matplotlib>=3.10.8",
    "numpy==1.26.4",
    "opencv-python==4.10.0.84",
    "pandas==2.2.2",
    "pillow==10.4.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "scikit-learn==1.5.1",
    "scipy==1.14.0",
    "seaborn==0.13.2",
    "tensorflow>=2.18.1",
]
